Moved the Sequencer to StepSequencer.tsx which musicArea.tsx calls
